performance_run/requirements.txt
tensorflow-gpu
np_utils
numpy
matplotlib
nvidia_smi